+++ Pandora/Paradigm/Inventory/Ability/Viewable.hs
@@ -3,4 +3,4 @@ module Pandora.Paradigm.Inventory.Ability.Viewable where
 
 class Viewable i where
 	type family Viewing i e r :: *
-	view_ :: Viewing i e r
+	view :: Viewing i e r